requirements.txt
streamlit>=1.28
tensorflow>=2.14
numpy>=1.24
pillow>=9.5
